requirements-lint.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-lint.in
#
appdirs==1.4.4            # via black
attrs==19.3.0             # via flake8-bugbear, flake8-eradicate, flake8-implicit-str-concat
black==20.8b1             # via -r requirements-lint.in
click==7.1.2              # via black
eradicate==1.0            # via flake8-eradicate
flake8-assert-msg==1.1.1  # via -r requirements-lint.in
flake8-bugbear==20.1.4    # via -r requirements-lint.in
flake8-builtins==1.5.3    # via -r requirements-lint.in
flake8-comprehensions==3.2.3  # via -r requirements-lint.in
flake8-eradicate==0.4.0   # via -r requirements-lint.in
flake8-executable==2.0.4  # via -r requirements-lint.in
flake8-implicit-str-concat==0.1.0  # via -r requirements-lint.in
flake8-isort==4.0.0       # via -r requirements-lint.in
flake8-plugin-utils==1.3.1  # via flake8-pytest-style
flake8-pytest-style==1.3.0  # via -r requirements-lint.in
flake8-raise==0.0.5       # via -r requirements-lint.in
flake8-string-format==0.3.0  # via -r requirements-lint.in
flake8==3.8.3             # via -r requirements-lint.in, flake8-assert-msg, flake8-bugbear, flake8-builtins, flake8-comprehensions, flake8-eradicate, flake8-executable, flake8-isort, flake8-raise, flake8-string-format
isort==5.5.2              # via flake8-isort
mccabe==0.6.1             # via flake8
mypy-extensions==0.4.3    # via black, mypy
mypy==0.782               # via -r requirements-lint.in
pathspec==0.8.0           # via black
pycodestyle==2.6.0        # via flake8
pydocstyle==5.1.1         # via -r requirements-lint.in
pyflakes==2.2.0           # via flake8
regex==2020.7.14          # via black
snowballstemmer==2.0.0    # via pydocstyle
testfixtures==6.14.2      # via flake8-isort
toml==0.10.1              # via black
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.3  # via black, mypy
